+++ packages/patternfly-4/react-core/src/components/BackgroundImage/BackgroundImage.js
@@ -75,22 +75,21 @@ const BackgroundImage = ({ className, src, ...props }) => {
 
   return (
     <div className={css(styles.backgroundImage, bgStyles.bgOverrides, className)}>
-      <svg xmlns="http://www.w3.org/2000/svg" className="svg-filter">
-        <filter id="image_overlay" width="100%">
+      <svg xmlns="http://www.w3.org/2000/svg" className="pf-c-background-image__filter" width="0" height="0">
+        <filter id="image_overlay" width="">
           <feColorMatrix type="matrix"
             values="1 0 0 0 0
-                  1 0 0 0 0
-                  1 0 0 0 0
-                  0 0 0 1 0" />
-    
-        <feComponentTransfer colorInterpolationFilters="sRGB" result="duotone">
-          <feFuncR type="table" tableValues="0.086274509803922 0.43921568627451"></feFuncR>
-          <feFuncG type="table" tableValues="0.086274509803922 0.43921568627451"></feFuncG>
-          <feFuncB type="table" tableValues="0.086274509803922 0.43921568627451"></feFuncB>
-          <feFuncA type="table" tableValues="0 1"></feFuncA>
-        </feComponentTransfer> 
-      </filter>
-    </svg>
+            1 0 0 0 0
+            1 0 0 0 0
+            0 0 0 1 0" />
+          <feComponentTransfer colorInterpolationFilters="sRGB" result="duotone">
+            <feFuncR type="table" tableValues="0.086274509803922 0.43921568627451"></feFuncR>
+            <feFuncG type="table" tableValues="0.086274509803922 0.43921568627451"></feFuncG>
+            <feFuncB type="table" tableValues="0.086274509803922 0.43921568627451"></feFuncB>
+            <feFuncA type="table" tableValues="0 1"></feFuncA>
+          </feComponentTransfer>
+        </filter>
+      </svg>
     </div>
   );
 };