+++ packages/patternfly-4/react-core/src/components/Checkbox/Checkbox.tsx
@@ -15,7 +15,7 @@ export interface CheckboxProps
   isChecked?: boolean;
   checked?: boolean;
   /** A callback for when the Checkbox selection changes. */
-  onChange?(checked: boolean, event: React.FormEvent<HTMLInputElement>): void;
+  onChange?: (checked: boolean, event: React.FormEvent<HTMLInputElement>) => void;
   /** Label text of the checkbox. */
   label?: React.ReactNode;
   /** Id of the checkbox. */
@@ -38,15 +38,15 @@ export class Checkbox extends React.Component<CheckboxProps> {
     super(props);
   }
 
-  static defaultProps = defaultProps;
+  private static defaultProps = defaultProps;
 
-  handleChange = (event: React.FormEvent<HTMLInputElement>): void => {
+  private handleChange = (event: React.FormEvent<HTMLInputElement>): void => {
     if (this.props.onChange) {
       this.props.onChange(event.currentTarget.checked, event);
     }
   };
 
-  render() {
+  public render() {
     const {
       'aria-label': ariaLabel,
       className,