+++ packages/patternfly-4/react-core/src/components/Wizard/WizardToggle.tsx
@@ -41,11 +41,11 @@ export const WizardToggle: React.FunctionComponent<WizardToggleProps> = ({
       activeStepName = steps[i].name;
       break;
     } else if (steps[i].steps) {
-      for (let j = 0; j < steps[i].steps!.length; j++) {
-        if ((activeStep.id && steps[i].steps![j].id === activeStep.id) || steps[i].steps![j].name === activeStep.name) {
+      for (const step of steps[i].steps!) {
+        if ((activeStep.id && step.id === activeStep.id) || step.name === activeStep.name) {
           activeStepIndex = i + 1;
           activeStepName = steps[i].name;
-          activeStepSubName = steps[i].steps![j].name;
+          activeStepSubName = step.name;
           break;
         }
       }