+++ packages/patternfly-4/react-integration/demo-app-ts/src/components/demos/AvatarDemo/AvatarDemo.tsx
@@ -1,0 +1,18 @@
+import { Avatar, AvatarProps } from '@patternfly/react-core';
+import React, { Component } from 'react';
+import avatarImg from '../../../assets/images/imgAvatar.svg';
+export class AvatarDemo extends Component {
+  myAvatarProps: AvatarProps = {
+    className: "MyClassName",
+    src: avatarImg,
+    alt: 'avatar'
+  }
+  
+  render() {
+    return (
+        <Avatar src={this.myAvatarProps.src} alt={this.myAvatarProps.alt} />
+    );
+  }
+}
+
+export default AvatarDemo;