+++ packages/patternfly-3/react-console/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.10.39 (2019-05-07)
+
+
+### Bug Fixes
+
+* **React Icons documentation:** fix React Icons documentation link to all Patternfly React Icons. ([#1868](https://github.com/patternfly/patternfly-react/issues/1868)) ([54d5404](https://github.com/patternfly/patternfly-react/commit/54d5404)), closes [#1841](https://github.com/patternfly/patternfly-react/issues/1841)
+
+
+
+
+
+## 1.10.38 (2019-05-03)
+
+
+### Bug Fixes
+
+* **ClickboardCopy:** update imports and clean up ([#1876](https://github.com/patternfly/patternfly-react/issues/1876)) ([7a85bfb](https://github.com/patternfly/patternfly-react/commit/7a85bfb))
+
+
+
+
+
 ## 1.10.37 (2019-04-15)