+++ packages/patternfly-4/react-core/CHANGELOG.md
@@ -3,6 +3,112 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.16.17](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.16.16...@patternfly/react-core@3.16.17) (2019-05-08)
+
+**Note:** Version bump only for package @patternfly/react-core
+
+
+
+
+
+## [3.16.16](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.16.15...@patternfly/react-core@3.16.16) (2019-05-07)
+
+**Note:** Version bump only for package @patternfly/react-core
+
+
+
+
+
+## [3.16.15](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.16.14...@patternfly/react-core@3.16.15) (2019-05-07)
+
+
+### Bug Fixes
+
+* **Dropdown:** Send event when clicking on dropdown toggle ([#1667](https://github.com/patternfly/patternfly-react/issues/1667)) ([fec1d41](https://github.com/patternfly/patternfly-react/commit/fec1d41))
+
+
+
+
+
+## [3.16.14](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.16.13...@patternfly/react-core@3.16.14) (2019-05-06)
+
+
+### Bug Fixes
+
+* **BackgroundImage:** fixed background image spacing issue  ([#1923](https://github.com/patternfly/patternfly-react/issues/1923)) ([ce38bae](https://github.com/patternfly/patternfly-react/commit/ce38bae))
+
+
+
+
+
+## [3.16.13](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.16.12...@patternfly/react-core@3.16.13) (2019-05-06)
+
+
+### Bug Fixes
+
+* **popover:** Enable highlighting and selecting text in popover. ([#1757](https://github.com/patternfly/patternfly-react/issues/1757)) ([70be06f](https://github.com/patternfly/patternfly-react/commit/70be06f))
+
+
+
+
+
+## [3.16.12](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.16.11...@patternfly/react-core@3.16.12) (2019-05-06)
+
+**Note:** Version bump only for package @patternfly/react-core
+
+
+
+
+
+## [3.16.11](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.16.10...@patternfly/react-core@3.16.11) (2019-05-03)
+
+
+### Bug Fixes
+
+* **dropitems:** updates to drop down item and skip to content api ([#1924](https://github.com/patternfly/patternfly-react/issues/1924)) ([a21f5fb](https://github.com/patternfly/patternfly-react/commit/a21f5fb))
+
+
+
+
+
+## [3.16.10](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.16.9...@patternfly/react-core@3.16.10) (2019-05-01)
+
+
+### Bug Fixes
+
+* **styles:** update layout styles for layouts ([#1906](https://github.com/patternfly/patternfly-react/issues/1906)) ([22e74ac](https://github.com/patternfly/patternfly-react/commit/22e74ac))
+
+
+
+
+
+## [3.16.9](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.16.8...@patternfly/react-core@3.16.9) (2019-05-01)
+
+**Note:** Version bump only for package @patternfly/react-core
+
+
+
+
+
+## [3.16.8](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.16.7...@patternfly/react-core@3.16.8) (2019-05-01)
+
+**Note:** Version bump only for package @patternfly/react-core
+
+
+
+
+
+## [3.16.7](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.16.6...@patternfly/react-core@3.16.7) (2019-05-01)
+
+
+### Bug Fixes
+
+* **ClickboardCopy:** update imports and clean up ([#1876](https://github.com/patternfly/patternfly-react/issues/1876)) ([7a85bfb](https://github.com/patternfly/patternfly-react/commit/7a85bfb))
+
+
+
+
+
 ## [3.16.6](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.16.5...@patternfly/react-core@3.16.6) (2019-04-30)