+++ packages/patternfly-4/react-integration/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.0.1 (2019-05-08)
+
+
+### Bug Fixes
+
+* **BackgroundImage:** fixed background image spacing issue  ([#1923](https://github.com/patternfly/patternfly-react/issues/1923)) ([ce38bae](https://github.com/patternfly/patternfly-react/commit/ce38bae))
+
+
+
+
+
 ## [1.0.14](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-integration@1.0.13...@patternfly/react-integration@1.0.14) (2019-02-01)
 
 **Note:** Version bump only for package @patternfly/react-integration